move black hole to x/common
