--- app/utils/utils_test.go
@@ -1,15 +1,0 @@
-package utils
-
-import (
-	"fmt"
-	"testing"
-
-	"github.com/stretchr/testify/require"
-)
-
-func TestBlackHoleAddress(t *testing.T) {
-	addr := BlackHoleAddress()
-	a := addr.String()
-	fmt.Println(a)
-	require.Equal(t, addr.String(), "okexchain1qqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqupa6dx")
-}
